# docs/api/swagger.yaml
openapi: 3.1.0
info:
  title: PowerChain API
  version: 1.0.0
  description: >-
    Versioned APIs for the PowerChain Digital Energy Operating System,
    including identity, AI, Proof of Energy, tokenization, cross-chain PWRC,
    marketplace settlement, smart meters, carbon, telemetry and platform configuration.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /bridge/pwrc/transfer:
    post:
      tags: [Bridge]
      summary: Prepare a signature-required PWRC bridge transfer
      description: Solana PWRC circulating supply must never exceed native PWRC locked for the Solana integration.
      responses:
        "201": { description: Transfer prepared }
        "400": { description: Validation error }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: powerchain_session